+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-28T05:20:54.122Z",
+    "2026-05-28T06:00:16.795Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {