+++ packages/orq-rc/RELEASES.md
@@ -9499,3 +9499,13 @@ Based on:
 - [typescript v4.10.0-rc.47] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.47] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.47 - packages/orq-rc
+
+## 2026-05-28 06:03:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.48] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.48] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.48 - packages/orq-rc